Add javadoc to FulfilledProcessor methods
